Split apart text-only logging and binary logging.
Text-only logging does not depend on Span, but binary logging (LogByteSpan)
does.  And Span.h depends on text-only logging.

We were manually breaking the cycle caused by having both logging types in
CHIPLogging.h by not including Span.h from CHIPLogging.h and forward-declaring
Span instead.  This fixes things so that we have a TextOnlyLogging.h (which is
used by Span.h, via CodeUtils.h, and does not use Span at all), and a
BinaryLogging.h which uses Span and includes Span.h.
