+++ src/lib/support/BUILD.gn
@@ -140,8 +140,11 @@ static_library("support") {
     "Variant.h",
     "ZclString.cpp",
     "ZclString.h",
-    "logging/CHIPLogging.cpp",
+    "logging/BinaryLogging.cpp",
+    "logging/BinaryLogging.h",
     "logging/CHIPLogging.h",
+    "logging/TextOnlyLogging.cpp",
+    "logging/TextOnlyLogging.h",
     "verhoeff/Verhoeff.cpp",
     "verhoeff/Verhoeff.h",
     "verhoeff/Verhoeff10.cpp",
