+++ src/lib/support/logging/TextOnlyLogging.cpp
@@ -1,6 +1,5 @@
 /*
- *
- *    Copyright (c) 2020 Project CHIP Authors
+ *    Copyright (c) 2020-2023 Project CHIP Authors
  *    Copyright (c) 2013-2017 Nest Labs, Inc.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,12 +22,11 @@
  *
  */
 
-#include "CHIPLogging.h"
+#include "TextOnlyLogging.h"
 
 #include <lib/core/CHIPCore.h>
 #include <lib/support/CHIPMem.h>
 #include <lib/support/CodeUtils.h>
-#include <lib/support/Span.h>
 
 #include <platform/logging/LogV.h>
 
@@ -195,37 +193,6 @@ void Log(uint8_t module, uint8_t category, const char * msg, ...)
     va_start(v, msg);
     LogV(module, category, msg, v);
     va_end(v);
-}
-
-void LogByteSpan(uint8_t module, uint8_t category, const chip::ByteSpan & span)
-{
-    // Maximum number of characters needed to print 8 byte buffer including formatting (0x)
-    // 8 bytes * (2 nibbles per byte + 4 character for ", 0x") + null termination.
-    // Rounding up to 50 bytes.
-    char output[50];
-    size_t offset = 0;
-    for (unsigned int i = 0; i < span.size(); i++)
-    {
-        if (i % 8 == 0 && offset != 0)
-        {
-            Log(module, category, "%s", output);
-            offset = 0;
-        }
-        int result = snprintf(&output[offset], sizeof(output) - offset, "0x%02x, ", (unsigned char) span.data()[i]);
-        if (result > 0)
-        {
-            offset += static_cast<size_t>(result);
-        }
-        else
-        {
-            Log(module, chip::Logging::kLogCategory_Error, "Failed to print ByteSpan buffer");
-            return;
-        }
-    }
-    if (offset != 0)
-    {
-        Log(module, category, "%s", output);
-    }
 }
 
 void LogV(uint8_t module, uint8_t category, const char * msg, va_list args)
